backend/news-scrapper/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
dnspython==2.6.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
idna==3.6
jieba3k==0.35.1
joblib==1.3.2
lxml==5.1.0
newspaper3k==0.2.8
nltk==3.8.1
pillow==10.2.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
tinysegmenter==0.3
tldextract==5.1.1
tqdm==4.66.2
urllib3==2.2.1
